backend/pyproject.toml
[tool.poetry]
name = "fitgpt-backend"
version = "0.1.0"
description = "Backend for FitGPT application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
flask = "^2.3.3"
werkzeug = "^2.3.7"
elasticsearch = "^8.13.1"
langchain = "^0.1.8"
langchain-core = ">=0.1.27,<0.2.0"
langchain-community = "^0.0.27"
langchain-elasticsearch = "0.2.2"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"
flask-login = "^0.6.2"
oauthlib = "^3.2.2"
requests-oauthlib = "^1.3.1"
google-auth = "^2.25.1"
google-auth-oauthlib = "^1.2.0"
huggingface-hub = "^0.19.4"
transformers = "^4.35.2"
sentence-transformers = "^2.2.2"
fuzzywuzzy = "^0.18.0"
jinja2 = "^3.1.2"
python-Levenshtein = "^0.23.0"
gunicorn = "^21.2.0"
mistralai = "^0.1.8"
langchain-mistralai = "^0.0.5"
wheel = "^0.41.3"
scikit-learn = "^1.2.2"
firebase-admin = "^6.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 